refactor: improved post list page
#32
