+++ src/app/post/page.tsx
@@ -1,42 +1,22 @@
 'use client';
 
-import { useQuery } from '@tanstack/react-query';
-import { useSearchParams } from 'next/navigation';
-import React, { useState } from 'react';
+import React from 'react';
 
-import { fetchPosts } from '@/app/api/post';
 import ListFilter from '@/app/components/Post/ListFilter';
 import ListTable from '@/app/components/Post/ListTable';
 import Pagination from '@/app/components/Post/Pagination';
+import usePostList from '@/app/hooks/post/usePostList';
 import useReplaceSearchParams from '@/app/hooks/post/useReplaceSearchParams';
-import { PostsReq, PostsRes, PostType } from '@/app/types/post';
+import useSearchFilters from '@/app/hooks/post/useSearchFilters';
 
 export default function PostList() {
-  const searchParams = useSearchParams();
-
-  const [searchFilters, setSearchFilters] = useState<PostsReq>({
-    type: (searchParams.get('type') as PostType) || '',
-    q: searchParams.get('q') || '',
-    page: Number(searchParams.get('page') || 1),
-  });
-
+  const { searchFilters, setSearchFilters } = useSearchFilters();
   const {
     data: { data: posts, meta },
     isLoading,
     isError,
     isFetching,
-  } = useQuery({
-    queryKey: ['posts', searchFilters],
-    queryFn: () => fetchPosts(searchFilters),
-    initialData: {
-      data: [],
-      meta: {
-        current_page: 1,
-        last_page: 1,
-      },
-      links: {},
-    } as unknown as PostsRes,
-  });
+  } = usePostiList({ searchFilters });
 
   useReplaceSearchParams({ searchFilters });
